daemon: extract APILimiterObserver implementation to separate package
This commit extracts the APILimiterObserver implementation from the
daemon package to a separate one, to allow for its reuse.
